🚑 fix: Ignore files without extensions when generating translations
Previously, the `FolderGenerateTranslations` class would attempt to generate translations for all files in a given folder, regardless of whether or not they had a file extension. This resulted in errors and unnecessary processing. With this commit, we now skip files without file extensions, denoted by a lack of a period followed by one or more letters at the end of the filename.