✨ feat: Update default model for `send` method in `Api.ts`
This commit updates the `send` method in `Api.ts` to remove the model parameter and use a default value of `"gpt-3.5-turbo-0301"`. The model parameter was previously used to specify the OpenAI language model to use for completing chat messages, but now all messages will be completed with the same model.